missed type and size of fb in I_WIPE_MeltScreen

+++ doom64/i_main.c
@@ -829,13 +829,13 @@ void I_MoveDisplay(int x,int y) // 80006790
 
 void I_WIPE_MeltScreen(void) // 80006964
 {
-    u32 *fb;
+    u16 *fb;
     int y1;
     int tpos;
     int yscroll;
     int height;
 
-    fb = Z_Malloc((SCREEN_WD*SCREEN_HT)*sizeof(u32), PU_STATIC, NULL);
+    fb = Z_Malloc((SCREEN_WD*SCREEN_HT)*sizeof(u16), PU_STATIC, NULL);
 
     I_GetScreenGrab();
     D_memcpy(&cfb[vid_side][0], &cfb[vid_side ^ 1][0], (SCREEN_WD*SCREEN_HT)*sizeof(u16));
